* PreferencesController.[h|m] new preference mpdTimeout, and add some           sane default values for it, and for mpdHost and mpdPort         * CollectionController.[h|m] fix some warnings and code cleanup         * Info-gnustep.plist general bumps from PC
git-svn-id: svn+ssh://svn.savannah.nongnu.org/gap/trunk@1717 d25b4930-3292-e311-b074-0010c6de5da8

+++ user-apps/MPDCon/CollectionController.h
@@ -73,8 +73,8 @@
 // Browser delegate Methods
 - (int)browser:(NSBrowser *)sender numberOfRowsInColumn:(int)column;
 - (void)browser:(NSBrowser *)sender willDisplayCell:(id)cell
-          atRow:(int)row
-         column:(int)column;
+          atRow:(NSInteger)row
+         column:(NSInteger)column;
 - (NSString *)browser:(NSBrowser *)sender
         titleOfColumn:(int)column;
